+Small refactor +Minor bug fixes +New file
Did a great refactor with printTopCoordinates def. Also corrected some formatting bugs and created the boardModule to keep code structure more organized.
